requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
boto3==1.42.9
botocore==1.42.9
celery==5.6.0
channels==4.3.2
channels_redis==4.3.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
Django==5.2.9
django-allauth==65.13.1
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-filter==25.2
django-rest-auth==0.9.5
django-storages==1.14.6
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
jmespath==1.0.1
joblib==1.5.2
kombu==5.6.1
msgpack==1.1.2
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sqlparse==0.5.4
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
